resouces_moving_system refactoring

+++ Turret_1/code/map_structures/buildings/building_types/factories/factory_types/rocket_factory.cpp
@@ -42,7 +42,7 @@ void RocketFactory::produceResource()
 }
 
 
-bool RocketFactory::canAccept(const int resType) const
+bool RocketFactory::canAccept(const uint16_t resType) const
 {
 	if (resType == RES_IRON || resType == RES_COPPER || resType == RES_SILICON || resType == RES_SULFUR)
 	{
